always keep a minimum of X backup

+++ lib/backup.rb
@@ -40,10 +40,12 @@ private
 	# @todo This deserves its own class
 	def deleteOldBackup
 		backupToRemove = []
+		numberOfBackups = 0
 		Message.searchingOldBackups
 		Dir.foreach(@generalConfig.get 'path') do |file|
 			next if file == "."
 			next if file == ".."
+			numberOfBackups += 1
 			today = Date.today
 			filedate = Date.new file[0, 4].to_i, file[4, 2].to_i, file[6, 2].to_i
 			x = today - filedate
@@ -56,13 +58,18 @@ private
 			Message.noBackupsToRemove
 			@log.debug "No old backup to remove"
 		else
+			backupToRemove.sort!
+			numberOfBackupToKeep = @generalConfig.getInt('savesToKeep')
 			backupToRemove.each do |file|
-				aDirectory = File.join(@generalConfig.get('path'), file)
-				@log.info "Removing backup #{aDirectory}"
-				begin
-					FileUtils.remove_dir aDirectory, true
-				rescue
-					@log.warn "Old backup <#{aDirectory}> maybe not removed"
+				if numberOfBackups > numberOfBackupToKeep
+					aDirectory = File.join(@generalConfig.get('path'), file)
+					@log.info "Removing backup #{aDirectory}"
+					begin
+						FileUtils.remove_dir aDirectory, true
+					rescue
+						@log.warn "Old backup <#{aDirectory}> maybe not removed"
+					end
+					numberOfBackups -= 1
 				end
 			end
 			Message.endOfRemovingOldBackups
